Created folder for figures
Figures are often helpful for explaining things, and if we have a lot of them in the top level directory then it will become difficult to parse. Therefore I've made a folder to hold figures. In it I'm creating a file to keep track of what figures it contains and where they are in the book.