Delete Chapter_Version_Control.md
As agreed during the 19/12/18 zoom discussion delete the barely-started version of the version control chapter from the master branch. Once it is in a somewhat reasonable shape it will be re-merge in via pull request  #88. After that a new pull request will be opened and subsequent edits will be made. Sarah agreed to help out with resolving any merge conflicts arising from this.