Change advice from git commit -m
Previously in this chapter readers instructed to provide commit messages via git commit -m, but this only allows short messages to be provided. Change this to advise readers to simply use 'git commit' which allows a short message to be separated from the main body. This follows from Srah's review.