Create open_science.md
This chapter has already had materials dumped in and some editing done. This was done on my (r-j-arnold)'s original fork which had to be deleted because it was detatched from the alan turing version when that repo was made public.

Transfer the work from that form to the re-created fork.