MASSIVE COMMIT
This commit upgrades the jupyter-book to the most recent version.
It moves the content folder from /book to /book/website.

The command that was run is as follows:

jupyter-book upgrade /Users/sgibson/source/github/alan-turing-institute/the-turing-way/book/website/

Hopefully this will resolve the problems we've had with failed
Travis builds of the book.