VCS: Separate existing content on version control into sub-chapter
- This takes the main content of the existing version control chapter and places it into a file
 // subchapter on 'general' version control. The introduction (summary, how version control
is helpful) and the resources (checklists, what to learn next, further reading, bibliography) are
separated and will be placed into an index file (at the toplevel of the version control chapter)
and a resources file (as the last subchapter in the version control chapter). This structure
mimics the structure of e.g. the Open Research chapter. It helps to shorten the version control
chapter into more stand-alone units, paves the way to include a 'version control for data'
subchapter, and creates a wonderfully structured (topic - problem - solution - how to - good
practice) standalone intro to Git and GitHub