VCS: Draft index file for version control chapter
Place the sections 'summary' and 'how version control is helpful' sections of the
previous version control chapter into an index file that serves to be the top level
chapter // landing page of the version control chapter (mimicing the layout of the
open research chapter).

Add a draft of a table of contents to navigate through the chapter.