fixed typos, improved some wording
To thank you for providing this super-useful list, I went through and copy-edited it. (I hope that's helpful and not annoying!)

One thing I noticed is a lot of redundancy. In many cases, I got tired of editing the same phrase over and over and just left it alone. :-)

You might consider improving the maintainability of this document by leaving out some of the repeated points and referring the reader to other sections, as you did here, for example:
> Share the link of the gallery where the pre-recorded lightning talks are posted along with the shared notes from the session (see details in (Program (before the event))