Tools: Sketch a short helper to find missed acknowledgements
This script queries the git history of a repository and checks
the .allcontributorsrc file using fuzzy matching for any
not yet mentioned committer to the history. Allows to specify
a commit-ish from where to start the query. Reports the author
name of potentially unacknowledged contributors and a summary
of their commits.