Proposed solution to #1187.
This solution provides a Dockerfile to build a local container from an
ubuntu 18.04 Bionic Beaver base image. By prividing a similar
environment to the one in the CI (`.github/workflows/ci.yml`), we are
more likely to be able to build the book successfully on a local machine
for testing.