Update rdm-spreadsheets.md
Hi team! Whilst reading this chapter online, I noticed the list of ways to make a dataset look tidy appeared formatted strangely, and figured it had to do with the | symbol. I have also removed the "3." from the only title that seems to be numbered. I hope this is okay!