right align text using <div dir="rtl"></div> tag
Arabic is read from right to left but it's a bit challenging to change the direction of the text in markdown without HTML tags, so it needs to be aligned to the right. Tags `ltr` need to be used for every paragraph or line. It can't be applied to all the text at once.