Fix multiple hyperlinks
There are two immediate issues with the hyperlinks in this file: 
* line 48 has a broken link to "Open Research";
* line 141 has an external link to Vimeo that does not render as a hyperlink on Netlify.

There is also a meta-issue or question around how to style links. This is not broken, but is inconsistent with the rest of the page and if I'm not wrong, the other pages in the book. On this page there are hyperlinks styled as footnotes, for example on lines 80-81, where the external URL is aliased and then the alias is used as a link. The remaining hyperlinks use Markdown standard inline hyperlinks syntax [alias](hyperlink).