Update ELNs LIMS bullet point
I want of make it clear that there exists a seperate type of software, LIMS, which has a different domain of responsibility from an ELN and that ELNs somtimes contain LIMS functionality.
Integration of ELNs and LIMS is very useful but LIMS functionality, where it exists in ELNs, can be limited in comparison to dedicated LIMS solution.
If you have a big high throughput lab you may need to find a seperate dedicated LIMS in addition to your ELN