Homepage: avoid open source qualifier for The Turing Way
Alternative proposal for #3705 if the concept of « open source » is assumed
strictly related to software.

Context for origin of the proposal: https://github.com/the-turing-way/the-turing-way/pull/3705#issuecomment-2161237778

Problematic situation where « open source » is explained and usually assumed
in relation to software, but used to qualify the Turing Way handbook in
an unclear manner as it is a non-software artefact. This fuels confusion
about the concept.

This is coming next to debates on the meaning of open source beyond
software, which may be applied to digital resources whose sources are provided.
Questioning regarding the fact open source should be used only for software,
PR for this case.

Need of an appropriate use of the terminology through the resource and
provide explanation that would reflect this understanding.