Glossary: Open Source (Re)Definition
fix #3653

Create an "open source" definition in the glossary, by renaming the current
"open source software" definition and by extending the concept beyond
software for digital resources with available sources.
Follow discussions over confusion surrounding the concept of open
source, around a questioning on whether the Turing Way is an open source
artefact despite not being a software. An improvement to acknowledge that
open source seems to transcend software.

In relation with a research work (open-source-undefined.org) on the definition
and the meaning of open source to try to bring a form of state of the art on the
usage and understanding of the concept, a notion that remains vague.

Based on the definition of open source in the GitHub's glossary, which
already recognizes open source outside of software, with a
number of enhancements:
- For digital resources, "replace working materials" by "source files"
to get this source dimension as it can be used.
- Mention explicitly source code for software
- Remove "by anyone" statements to be neutral regarding OSI
  interpretation of open source
- For digital resources, duplicate use/modify/share instead of
  fork/modify/discuss/contribute originally from GitHub
- Add some terminology examples for open source resources.

The GitHub definition was edited after an open conversation including
workers from GitHub, initially around the place of collaboration in
the definition but which reach questioning regarding sources of blog
posts or other resources, leading to their conclusion as article title
« Open source, not just software anymore » (see bibliography).

This meaning of open source is in tension with the more conventional
usage of the term, which results in a kind of double meaning between an
open source limited to software or something concerning all types of
digital resources. This would be a sign of important limitations in
reference definitions of the field, notably the Open Source Definition.

We could be witnessing an evolution in the meaning of the concept of
open source, a better understanding of the concept sounds necessary.

According to this definition, The Turing Way is open source.

Bibliography:

- title: Definition of open source in GitHub's glossary
  link: https://docs.github.com/en/get-started/learning-about-github/github-glossary#open-source
- title: Open source, not just software anymore by Ben Balter, 2014
  link: https://ben.balter.com/2014/01/27/open-collaboration/
- title: Open source beyond software affirmation by Karl Fogel, 2014
  link: https://github.com/benbalter/benbalter.github.com/pull/98#issuecomment-33644426
- title: Editorial Defining Open Source, authored by Russ Nelson for the
  Technology Innovation Management Review, 2007
  link: https://timreview.ca/issue/2007/september
- title: Resource « What is open source ? » on opensource.com
  link: https://opensource.com/resources/what-open-source
- title: Developing Open Source Educational Resources for Machine
  Learning and Data Science, 2021
  link: https://doi.org/10.48550/arXiv.2107.14330
- title: What do you mean by "open-source educational materials"?, about
  section of the Journal of Open Source Education
  link: https://jose.theoj.org/about
- title: Can a novel be open source, although it does not have a
  source?, StackExchange
  link: https://opensource.stackexchange.com/questions/423/can-a-novel-be-open-source-although-it-does-not-have-a-source
- title: Definition of Free Cultural Works, 2006
  link: https://freedomdefined.org/Definition
- title: Definition of open source on Wikitionary (noun, adjective)
  link: https://en.wiktionary.org/wiki/open-source
- title: Open source page of Wikipedia
  link: https://en.wikipedia.org/wiki/Open_source
- title: Open Source Hardware Definition
  link: https://www.oshwa.org/definition/
- title: Measuring Openness in Open Source Hardware with the Open-o-Meter
  link: https://www.sciencedirect.com/science/article/pii/S2212827118312095