Create community growth & sustainability section
This section provides a guide that allows readers to develop a fully
evidenced "community roadmap" to help with the issues of both growing
and sustaining a community. It was developed by Open@RIT as part of
their work assisting academic open work projects. There may be some
overlap with the existing personas section.

Seeks to resolve #3785